### 2021-05-05 #3252 @sibirrer
Hi JOSS editors!

The paper is in the branch JOSS_paper (I somehow could not indicate this on the submission form/was rejected as not a clone-able project). I may have mis-clicked on the suggested editor form. I am happy with any editor.

For reviewers, when reading from bottom to top of the file, I can identify two lensing experts:
coljac
smsharma

and in addition those people with coding and astronomy expertise I noted (incomplete list): changhoonhahn
RobertRosca
Teuben
olebole
danhey
cmillion
KshitijAggarwal
yymao
HeloiseS

Let me know if you have questions and thank you very much for handling our submission!
Simon, in the name of the authors
