### 2021-05-05 30318ac
minor typo and style fixes, added acknowledgement
