+++ Paper/paper.md
@@ -75,7 +75,7 @@ affiliations:
    index: 1
  - name: SLAC National Accelerator Laboratory, Menlo Park, CA, 94025, USA
    index: 2
- - name: Department of Astronomy & Astrophysics, University of Chicago, Chicago, IL 606374, USA
+ - name: Department of Astronomy & Astrophysics, University of Chicago, Chicago, IL 60637, USA
    index: 3
  - name: Department of Physics and Astronomy, University of California, Los Angeles, CA 90095, USA
    index: 4 
@@ -131,12 +131,12 @@ Bottom row from right to left along the red arrow: The inverse problem is solved
 
 # Background
 
-Gravitational lensing displaces the observed positions and distorts the shapes of apparent objects on the sky due to intervening inhomogeneous matter along the line of sight. Strong gravitational lensing describes the regime where the background source, such as a galaxy or quasar, is lensed by a massive foreground object, such as another galaxy or cluster of galaxies, to produce multiple images of itself in a highly distorted manner. 
+Gravitational lensing displaces the observed positions and distorts the shapes of apparent objects on the sky due to intervening inhomogeneous matter along the line of sight. Strong gravitational lensing describes the regime where the background source, such as a galaxy or quasar, is lensed by a massive foreground object, such as another galaxy or cluster of galaxies, to produce multiple images of the source in a highly distorted manner. 
 The top row of \autoref{fig:example} illustrates such a process from the intrinsic galaxy to the data product at hand, including the lensing distortions, effects of the instrument, observational conditions and noise.
 
 Analyses of strong gravitational lensing have provided a wealth of key insights into cosmology and astrophysics.
 For example, relative time delays of multiply imaged variable sources provided precision measurements on the expansion rate of the Universe [@Wong:2020; @Shajib:2020strides; @Birrer:2020tdcosmoiv]. Small scale distortions in the lensing signal of resolved sources [@Vegetti:2012; @Hezaveh:2016; @Birrer:2017]
-and unresolved flux ratios [@Gilman:2020; @Hsueh:2020] constrain the nature of dark matter. Combined strong lensing and kinematic observables constrain the formation and evolution of galaxies [@Sonnenfeld:2015; @Shajib:2021slacs], and the lensing magnification effect provides an otherwise inaccessible angle on the early universe [@Zheng:2012; @Cava:2018].
+and unresolved flux ratios [@Gilman:2020; @Hsueh:2020] constrain the nature of dark matter. Combined strong lensing and kinematic observables constrain the formation and evolution of galaxies [@Sonnenfeld:2015; @Shajib:2021slacs], and the lensing magnification effect provides an otherwise inaccessible angle on the early Universe [@Zheng:2012; @Cava:2018].
 
 
 
@@ -144,15 +144,15 @@ and unresolved flux ratios [@Gilman:2020; @Hsueh:2020] constrain the nature of d
 # Statement of need
 
 Strong lensing studies have significantly enhanced, and sometimes challenged, our current fundamental understanding of the Universe.
-In the near future, with the onset of the next generation ground and space-based wide and deep astronomical imaging [Rubin, Roman, Euclid observatories; @LSST; @Roman; @Euclid] and interferometric [SKA; @SKA] surveys, the number of discovered lenses of different types will be growing by more than an order of magnitude [@Collett:2015; @OM10].
+In the near future, with the onset of the next-generation ground and space-based wide and deep astronomical imaging [Rubin, Roman, Euclid observatories; @LSST; @Roman; @Euclid] and interferometric [SKA; @SKA] surveys, the number of discovered lenses of different types will be growing by more than an order of magnitude [@Collett:2015; @OM10].
 Such large samples can provide unprecedented statistical precision to stress-test our current understanding and exploit discovery potential.
[…]
