### 2021-05-13 #3252 @danielskatz
👋 @coljac, @smsharma - I'm looking for two reviewers for this JOSS submission.  Are either or ideally both of you willing to review this?
