### 2021-05-13 #3283 @whedon
[REVIEW]: lenstronomy II: A gravitational lensing software ecosystem
**Submitting author:** @sibirrer (<a href="http://orcid.org/0000-0003-3195-5507">Simon Birrer</a>)
**Repository:** <a href="https://github.com/sibirrer/lenstronomy" target ="_blank">https://github.com/sibirrer/lenstronomy</a>
**Version:** 1.8.2_JOSS
