### 2021-05-14 #3283 @danielskatz
👋 @sibirrer - I forgot to ask before: Is this submission associated with a paper submitted to a AAS journal?
