### 2021-05-14 #3283 @sibirrer
> 👋 @sibirrer - I forgot to ask before: Is this submission associated with a paper submitted to a AAS journal?

@danielskatz No, this is a stand-alone submission. There was an original paper in 2018 (Physics of the Dark Universe) (without code review, highlighting the design feature).
