### 2021-05-14 #3283 @danielskatz
Thanks - it doesn't really matter in terms of the review process, but it would lead to a minor difference in the publication at the end - see https://blog.joss.theoj.org/2018/12/a-new-collaboration-with-aas-publishing for details
