### 2021-05-14 #3283 @sibirrer
I see, thanks for the pointer! Was not fully aware of this, but also think that a full-fledged paper is too much work in writing. I hear your argument of financial support through AAS submissions. Perhaps future papers that are responsible for implementing specific features in lenstronomy might be worth going through a JOSS review (it would only be specific modules of lenstronomy). This would both ensure a steady high quality of added features and a rewarding structure for both JOSS and the authors that implemented specific features (it's not just me at this point). It's just a thought but it would also also put more work on JOSS and more frequent reviews.
