### 2021-05-14 #3283 @danielskatz
Note that one of our (JOSS) sensitivities is in being sure that submissions are substantial units of work, in some equivalent to a science paper, so this might be an issue. In any case, the question has been answered for this submission :)
