+++ lenstronomy/LensModel/Profiles/cored_density_mst.py
@@ -48,7 +48,7 @@ class CoredDensityMST(LensProfileBase):
         kappa_ext = 1 - lambda_approx
         f_cored_density = self._profile.function(x, y, kappa_ext, r_core, center_x, center_y)
         f_ms = self._convergence.function(x, y, kappa_ext, center_x, center_y)
-        return (f_cored_density - f_ms)/kappa_ext
+        return (f_cored_density - f_ms)/lambda_approx
 
     def derivatives(self, x, y, lambda_approx, r_core, center_x=0, center_y=0):
         """
@@ -65,7 +65,7 @@ class CoredDensityMST(LensProfileBase):
         kappa_ext = 1 - lambda_approx
         f_x_cd, f_y_cd = self._profile.derivatives(x, y, kappa_ext, r_core, center_x, center_y)
         f_x_ms, f_y_ms = self._convergence.derivatives(x, y, kappa_ext, center_x, center_y)
-        return (f_x_cd - f_x_ms)/kappa_ext, (f_y_cd - f_y_ms)/kappa_ext
+        return (f_x_cd - f_x_ms)/lambda_approx, (f_y_cd - f_y_ms)/lambda_approx
 
     def hessian(self, x, y, lambda_approx, r_core, center_x=0, center_y=0):
         """
@@ -82,4 +82,4 @@ class CoredDensityMST(LensProfileBase):
         kappa_ext = 1 - lambda_approx
         f_xx_cd, f_yy_cd, f_xy_cd = self._profile.hessian(x, y, kappa_ext, r_core, center_x, center_y)
         f_xx_ms, f_yy_ms, f_xy_ms = self._convergence.hessian(x, y, kappa_ext, center_x, center_y)
-        return (f_xx_cd - f_xx_ms)/kappa_ext, (f_yy_cd - f_yy_ms)/kappa_ext, (f_xy_cd - f_xy_ms)/kappa_ext
+        return (f_xx_cd - f_xx_ms)/lambda_approx, (f_yy_cd - f_yy_ms)/lambda_approx, (f_xy_cd - f_xy_ms)/lambda_approx
