### 2021-05-20 c1852c6
update lens_model_plot to be consistent with the axis coordinates. Before: axis where given with (0,0) at the edge of the image. This is a convenient notation in case of rotated coordinate systems but is not desired for the illustration plot where the axis are fixed. Also added a feature in the critical curve computation to be offset from the center. This commit addresses issue #253
