+++ lenstronomy/Plots/lens_plot.py
@@ -18,15 +18,24 @@ export, __all__ = exporter()
 
 @export
 def lens_model_plot(ax, lensModel, kwargs_lens, numPix=500, deltaPix=0.01, sourcePos_x=0, sourcePos_y=0,
-                    point_source=False, with_caustics=False, with_convergence=True, coord_center_ra=0, coord_center_dec=0,
-                    coord_inverse=False, fast_caustic=False):
+                    point_source=False, with_caustics=False, with_convergence=True, coord_center_ra=0,
+                    coord_center_dec=0, coord_inverse=False, fast_caustic=False):
     """
     plots a lens model (convergence) and the critical curves and caustics
 
-    :param ax:
-    :param kwargs_lens:
-    :param numPix:
-    :param deltaPix:
+    :param ax: matplotlib axis instance
+    :param kwargs_lens: lens model keyword argument list
+    :param numPix: total nnumber of pixels (for convergence map)
+    :param deltaPix: width of pixel (total frame size is deltaPix x numPix)
+    :param sourcePos_x: float, x-position of point source (image positions computed by the lens equation)
+    :param sourcePos_y: float, y-position of point source (image positions computed by the lens equation)
+    :param point_source: bool, if True, illustrates and computes the image positions of the point source
+    :param with_caustics: bool, if True, illustrates the critical curve and caustics of the system
+    :param with_convergence: bool, if True, illustrates the convergence map
+    :param coord_center_ra: float, x-coordinate of the center of the frame
+    :param coord_center_dec: float, y-coordinate of the center of the frame
+    :param coord_inverse: bool, if True, inverts the x-coordinates to go from right-to-left
+     (effectively the RA definition)
     :param fast_caustic: boolean, if True, uses faster but less precise caustic calculation
      (might have troubles for the outer caustic (inner critical curve)
     :param with_convergence: boolean, if True, plots the convergence of the deflector
@@ -38,44 +47,60 @@ def lens_model_plot(ax, lensModel, kwargs_lens, numPix=500, deltaPix=0.01, sourc
     _coords = data
     _frame_size = numPix * deltaPix
     x_grid, y_grid = data.pixel_coordinates
+    ra0, dec0 = data.radec_at_xy_0
+    origin = [ra0, dec0]
+    if coord_inverse:
+        extent = [ra0, ra0 - _frame_size, dec0, dec0 + _frame_size]
+        delta_pix_x = - deltaPix
+    else:
+        extent = [ra0, ra0 + _frame_size, dec0, dec0 + _frame_size]
+        delta_pix_x = deltaPix
     lensModelExt = LensModelExtensions(lensModel)
     x_grid1d = util.image2array(x_grid)
     y_grid1d = util.image2array(y_grid)
     if with_convergence:
         kappa_result = lensModel.kappa(x_grid1d, y_grid1d, kwargs_lens)
         kappa_result = util.array2image(kappa_result)
-        im = ax.matshow(np.log10(kappa_result), origin='lower', extent=[0, _frame_size, 0, _frame_size], cmap='Greys',
+        im = ax.matshow(np.log10(kappa_result), origin='lower', extent=extent, cmap='Greys',
                         vmin=-1, vmax=1) #, cmap=self._cmap, vmin=v_min, vmax=v_max)
     if with_caustics is True:
         if fast_caustic:
-            ra_crit_list, dec_crit_list, ra_caustic_list, dec_caustic_list = lensModelExt.critical_curve_caustics(kwargs_lens, compute_window=_frame_size, grid_scale=deltaPix)
[…]
