### 2021-05-21 103cfff
outcommenting readme figure for readthedocs compilation
