### 2021-05-21 aa8b0d6
docstrings updated
