### 2021-05-23 03bb353
fixing test functions
