+++ test/test_LensModel/test_Profiles/test_cored_profile_mst.py
@@ -17,7 +17,8 @@ class TestMassSheet(object):
         self.profile1 = CoredDensityMST(profile_type='CORED_DENSITY')
         self.profile2 = CoredDensityMST(profile_type='CORED_DENSITY_2')
         self.profile3 = CoredDensityMST(profile_type='CORED_DENSITY_EXP')
-        self.kwargs_lens = {'lambda_approx': 1.1, 'r_core': 100, 'center_x': 0, 'center_y': 0}
+        self.profile4 = CoredDensityMST(profile_type='CORED_DENSITY_ULDM')
+        self.kwargs_lens = {'lambda_approx': 0.9, 'r_core': 100, 'center_x': 0, 'center_y': 0}
 
     def test_function(self):
         x = np.array([0.01, 1])
@@ -25,6 +26,10 @@ class TestMassSheet(object):
         f_ = self.profile1.function(x, y, **self.kwargs_lens)
         npt.assert_almost_equal(f_[0] - f_[1], 0, decimal=3)  # test to demand that the profile is (almost) zero
         f_ = self.profile2.function(x, y, **self.kwargs_lens)
+        npt.assert_almost_equal(f_[0] - f_[1], 0, decimal=3)  # test to demand that the profile is (almost) zero
+        f_ = self.profile3.function(x, y, **self.kwargs_lens)
+        npt.assert_almost_equal(f_[0] - f_[1], 0, decimal=3)  # test to demand that the profile is (almost) zero
+        f_ = self.profile4.function(x, y, **self.kwargs_lens)
         npt.assert_almost_equal(f_[0] - f_[1], 0, decimal=3)  # test to demand that the profile is (almost) zero
 
     def test_derivatives(self):
@@ -34,6 +39,10 @@ class TestMassSheet(object):
         npt.assert_almost_equal(f_x[0] - f_x[1], 0, decimal=3)  # test to demand that the profile is (almost) zero
         f_x, f_y = self.profile2.derivatives(x, y, **self.kwargs_lens)
         npt.assert_almost_equal(f_x[0] - f_x[1], 0, decimal=3)  # test to demand that the profile is (almost) zero
+        f_x, f_y = self.profile3.derivatives(x, y, **self.kwargs_lens)
+        npt.assert_almost_equal(f_x[0] - f_x[1], 0, decimal=3)  # test to demand that the profile is (almost) zero
+        f_x, f_y = self.profile4.derivatives(x, y, **self.kwargs_lens)
+        npt.assert_almost_equal(f_x[0] - f_x[1], 0, decimal=3)  # test to demand that the profile is (almost) zero
 
     def test_hessian(self):
         x = np.array([0.01, 1])
@@ -41,6 +50,10 @@ class TestMassSheet(object):
         f_xx, f_yy, f_xy = self.profile1.hessian(x, y, **self.kwargs_lens)
         npt.assert_almost_equal(f_xx, 0, decimal=3)  # test to demand that the profile is (almost) zero
         f_xx, f_yy, f_xy = self.profile2.hessian(x, y, **self.kwargs_lens)
+        npt.assert_almost_equal(f_xx, 0, decimal=3)  # test to demand that the profile is (almost) zero
+        f_xx, f_yy, f_xy = self.profile3.hessian(x, y, **self.kwargs_lens)
+        npt.assert_almost_equal(f_xx, 0, decimal=3)  # test to demand that the profile is (almost) zero
+        f_xx, f_yy, f_xy = self.profile4.hessian(x, y, **self.kwargs_lens)
         npt.assert_almost_equal(f_xx, 0, decimal=3)  # test to demand that the profile is (almost) zero
 
 
