+++ test/test_LensModel/test_Profiles/test_uldm.py
@@ -20,17 +20,17 @@ class TestUldm(object):
         r = np.sqrt(x**2 + y**2)
         kappa_0 = 0.1
         theta_c = 5
-        slope = 8.5
-        f_ = self.model.function(r, 0, kappa_0, theta_c, slope)
+        slope = 7.8
+        f_ = self.model.function(r, 0, kappa_0, theta_c, 0, 0, slope)
         delta = 0.0001
-        f_d = self.model.function(r + delta, 0, kappa_0, theta_c, slope)
+        f_d = self.model.function(r + delta, 0, kappa_0, theta_c, 0, 0, slope)
         f_x_num = (f_d - f_) / delta
-        f_x, _ = self.model.derivatives(r, 0, kappa_0, theta_c, slope)
+        f_x, _ = self.model.derivatives(r, 0, kappa_0, theta_c, 0, 0, slope)
         npt.assert_almost_equal(f_x_num, f_x, decimal=3)
         # Try MSD limit
         theta_c_large = 10
-        f_reference = self.model.function(0, 0, kappa_0, theta_c_large, slope)
-        f_large = self.model.function(r, 0, kappa_0, theta_c_large, slope)
+        f_reference = self.model.function(0, 0, kappa_0, theta_c_large, 0, 0, slope)
+        f_large = self.model.function(r, 0, kappa_0, theta_c_large, 0, 0, slope)
         f_MSD = 0.5* kappa_0 * r**2
         npt.assert_almost_equal(f_large - f_reference, f_MSD, decimal=3)
 
@@ -39,8 +39,8 @@ class TestUldm(object):
         y = 0.8
         r = np.sqrt(x**2 + y**2)
         kappa_0, theta_c = 0.2, 9 # Trying MSD limit
-        slope = 8.5
-        f_x, f_y = self.model.derivatives( x, y, kappa_0, theta_c, slope)
+        slope = 6.5
+        f_x, f_y = self.model.derivatives( x, y, kappa_0, theta_c, 0, 0, slope)
         alpha_MSD = kappa_0 * r
         npt.assert_almost_equal(f_x, alpha_MSD * x/r, decimal=3)
         npt.assert_almost_equal(f_y, alpha_MSD * y/r, decimal=3)
@@ -51,8 +51,8 @@ class TestUldm(object):
         r = np.sqrt(x**2 + y**2)
         kappa_0 = 0.12
         theta_c = 6
-        slope = 8.5
-        f_xx, f_yy, f_xy = self.model.hessian(x, y, kappa_0, theta_c, slope)
+        slope = 5.5
+        f_xx, f_yy, f_xy = self.model.hessian(x, y, kappa_0, theta_c, 0, 0, slope)
         kappa = 1./2 * (f_xx + f_yy)
         kappa_direct = self.model.kappa_r(r, kappa_0, theta_c, slope)
         npt.assert_almost_equal(kappa, kappa_direct, decimal=5)
@@ -63,7 +63,7 @@ class TestUldm(object):
         r = np.sqrt(x ** 2 + y ** 2)
         kappa_0 = 0.1
         theta_c = 7
-        slope = 8.5
+        slope = 4.5
         m3d = self.model.mass_3d(r, kappa_0, theta_c, slope)
         m3d_lens = self.model.mass_3d_lens(r, kappa_0, theta_c, slope)
         npt.assert_almost_equal(m3d, m3d_lens, decimal=8)
