### 2021-06-01 #3283 @coljac
OK, sorry it took so long to get this done.

This is in easy review in that Lenstronomy is a mature, feature-rich and widely used package. The scientific significance is not in doubt, and a lot of time and effort has been invested into the code, docs and examples. It shows. Congrats all round to the authors!

I have no issues with the paper.

@sbirrer, could you clarify these final issues for me:

- One of the dependencies, pypolychord, appears to have been yanked from pypi. So `pip install -r requirements.txt` fails for me. I installed from the github repo, but this may be an issue for other users.
- Where is running the tests documented? I didn't find the docs, but I noticed the tox.ini. I tried running the tests via tox, and get another dependency issue (related to the py27 environment). I did get all the tests to pass with py.test however, so that's all good.
- In the notebooks, did have to chase down one or two more dependencies, e.g. astrofunc (not on pypi, had to find the repo).

Barring the above, which are very minor, I'm happy to sign off.
