### 2021-06-01 e45b67b
tox testing updated to python3, colossus comparison without cache to allow for virtual environment, installation description with added tests. This PR addresses a comment of the JOSS review https://github.com/openjournals/joss-reviews/issues/3283
