### 2021-06-02 #3283 @smsharma
Apologies likewise for the delay on this!

`lenstronomy` is a mature package widely used in strong lensing analyses for modeling a variety of relevant physical systems (e.g., time-delay lenses, galaxy-galaxy lenses) and enabling downstream tasks such as Hubble constant measurement and dark matter substructure inference. The modular API additionally makes it easy to adapt for a given purpose as well as to contribute to the codebase. This has also enabled a small ecosystem of more specialized packages around `lenstronomy`. The package clearly demonstrates substantial scholarly effort.

I had no issues with installation as recommended in the documentation, and was able to go through the Jupyter notebook examples in the `lenstronomy_extensions` repository. The package ticks all the boxes at this stage, and I'm happy to sign off. Congrats to @sibirrer and all involved!
