### 2021-06-02 #3283 @sibirrer
Thank you @smsharma  and @coljac  for the very encouraging and positive feedback! @coljac I also updated the notebooks slightly, in particular the astrofunc dependency got removed (was a leftover as the functionalities are itself available in lenstronomy).

@danielskatz what are the next steps for me/us?
