### 2021-06-02 757a91d
Implementation of Jeans method SIDM profiles.
Todo:

- Tests
- Hessian
- Documentation
- Examples
