+++ lenstronomy/LensModel/Profiles/jeans_iso.py
@@ -1,0 +1,515 @@
+__author__ = 'jhodonnell'
+
+import numpy as np
+
+from scipy import integrate
+from scipy.interpolate import interp1d
+from scipy.optimize import root
+
+import warnings
+
+from lenstronomy.LensModel.Profiles.base_profile import LensProfileBase
+
+__all__ = ['SIDMJeans']
+
+# ## Analytic Solution
+#
+# As mentioned above, for x close to zero we have an exact solution. Below a
+# small threshold, we will approx y(x)/y'(x) with these solutions, above that
+# threshold we will use odeint.
+
+
+def _y_approx(x):
+    return np.log(1 - np.tanh(x/np.sqrt(6))**2)
+
+
+def _dy_approx(x):
+    return -np.sqrt(2/3)*np.tanh(x/np.sqrt(6))
+
+
+def _ddy_approx(x):
+    return -1/(3*np.cosh(x/np.sqrt(6)))
+
+
+class _JeansSolutionInterp:
+    '''
+    Solves the Jeans method differential equation, and boundary conditions
+    between the isothermal core and NFW exterior.
+    '''
+
+    def _solve_jeans_function(self):
+        '''
+        Solve the isothermal Jeans differential eqn
+        '''
+        def _jeans_function_integrand(y, x):
+            y, dy = y
+            ddy = -(np.exp(y) + 2*dy / x)
+            return np.array([dy, ddy])
+
+        xspan = np.logspace(-3, 2, 10000)
+        start_x = (xspan[0])
+        solution = integrate.odeint(
+            _jeans_function_integrand,
+            y0=[_y_approx(start_x), _dy_approx(start_x)],
+            t=xspan
+        )
+
+        self._y_start_x = start_x
+        self._xspan_y = xspan
+        self._y_interp = interp1d(
+            xspan, solution[:, 0], fill_value='extrapolate'
+        )
+        self._dy_interp = interp1d(
+            xspan, solution[:, 1], fill_value='extrapolate'
+        )
+
+    def _solve_mass(self):
+        '''
+        Compute the enclosed mass of the isothermal Jeans eqn
+        '''
+        def mass_integrand(y_, x):
+            return x*x*np.exp(self.y(x))
+
+        xs_mass = np.concatenate([[0.0], np.logspace(-3, 2, 9999)])
+        mass_integrated = integrate.odeint(mass_integrand, y0=[0], t=xs_mass).flatten()
+
+        self._xspan_mass = xs_mass
+        self._mass_interp = interp1d(
+            xs_mass, mass_integrated, fill_value='extrapolate'
+        )
+
+    def y(self, x):
+        '''
+        Give the (unitless) solution to the reduced Jeans function y(x)
+        '''
+        if not hasattr(self, '_y_interp'):
+            self._solve_jeans_function()
+
+        if isinstance(x, (int, float)):
+            if x < self._y_start_x:
+                return _y_approx(x)
+            else:
+                return self._y_interp(x)
+
+        output = np.empty_like(x)
+        output[x <= self._y_start_x] = _y_approx(x[x <= self._y_start_x])
+        output[x > self._y_start_x] = self._y_interp(x[x > self._y_start_x])
+        return output
+
+    def enclosed_mass_3d(self, x):
+        '''
+        Give the (unitless) mass enclosed within the sphere r < r0*x
+        '''
+        if not hasattr(self, '_mass_interp'):
+            self._solve_mass()
+
+        if isinstance(x, (int, float)):
+            if x < self._xspan_mass.min():
+                return 0.0
+            elif x > self._xspan_mass.max():
+                xmax = self._xspan_mass.max()
+                return (
+                    self._mass_interp(xmax)*(x/xmax)**2
+                )
[…]
