### 2021-06-04 8d8b711
default lower and upper bounds of EPL model adjusted. Previous bounds are meant to be in projected quantity
