+++ lenstronomy/LensModel/Profiles/epl.py
@@ -44,8 +44,8 @@ class EPL(LensProfileBase):
     scheme.
     """
     param_names = ['theta_E', 'gamma', 'e1', 'e2', 'center_x', 'center_y']
-    lower_limit_default = {'theta_E': 0, 'gamma': 0, 'e1': -0.5, 'e2': -0.5, 'center_x': -100, 'center_y': -100}
-    upper_limit_default = {'theta_E': 10, 'gamma': 2, 'e1': 0.5, 'e2': 0.5, 'center_x': 100, 'center_y': 100}
+    lower_limit_default = {'theta_E': 0, 'gamma': 1, 'e1': -0.5, 'e2': -0.5, 'center_x': -100, 'center_y': -100}
+    upper_limit_default = {'theta_E': 10, 'gamma': 3, 'e1': 0.5, 'e2': 0.5, 'center_x': 100, 'center_y': 100}
 
     def __init__(self):
         self.epl_major_axis = EPLMajorAxis()
