### 2021-06-04 c826e80
further adjustment of theta_E and gamma bounds to be the same among EPL, EPL_NUMBA, PEMD
