+++ lenstronomy/LensModel/Profiles/tnfw.py
@@ -4,8 +4,8 @@ __author__ = 'sibirrer'
 # the potential therefore is its integral
 
 import numpy as np
-import warnings
 from lenstronomy.LensModel.Profiles.base_profile import LensProfileBase
+from lenstronomy.LensModel.Profiles.nfw import NFW
 
 __all__ = ['TNFW']
 
@@ -17,6 +17,7 @@ class TNFW(LensProfileBase):
     relation are: R_200 = c * Rs
 
     """
+    profile_name = 'TNFW'
     param_names = ['Rs', 'alpha_Rs', 'r_trunc', 'center_x', 'center_y']
     lower_limit_default = {'Rs': 0, 'alpha_Rs': 0, 'r_trunc': 0, 'center_x': -100, 'center_y': -100}
     upper_limit_default = {'Rs': 100, 'alpha_Rs': 10, 'r_trunc': 100, 'center_x': 100, 'center_y': 100}
@@ -24,7 +25,6 @@ class TNFW(LensProfileBase):
     def __init__(self):
         """
 
-        :param interpol: bool, if True, interpolates the functions F(), g() and h()
         """
         self._s = 0.001
         super(LensProfileBase, self).__init__()
@@ -36,13 +36,12 @@ class TNFW(LensProfileBase):
         :param y: angular position
         :param Rs: angular turn over point
         :param alpha_Rs: deflection at Rs
+        :param r_trunc: truncation radius
         :param center_x: center of halo
         :param center_y: center of halo
-        :return:
+        :return: lensing potential
         """
-        rho0_input = self._alpha2rho0(alpha_Rs=alpha_Rs, Rs=Rs)
-        #if Rs < 0.0001:
-        #    Rs = 0.0001
+        rho0_input = self.alpha2rho0(alpha_Rs=alpha_Rs, Rs=Rs)
         x_ = x - center_x
         y_ = y - center_y
         R = np.sqrt(x_ ** 2 + y_ ** 2)
@@ -51,7 +50,7 @@ class TNFW(LensProfileBase):
 
         return f_
 
-    def L(self, x, tau):
+    def _L(self, x, tau):
         """
         Logarithm that appears frequently
         :param x: r/Rs
@@ -88,7 +87,7 @@ class TNFW(LensProfileBase):
             else:
                 return (x ** 2 - 1) ** -.5 * np.arctan((x ** 2 - 1) ** .5)
 
-    def derivatives(self, x, y, Rs=None, alpha_Rs=None, r_trunc=None, center_x=0, center_y=0):
+    def derivatives(self, x, y, Rs, alpha_Rs, r_trunc, center_x=0, center_y=0):
 
         """
         returns df/dx and df/dy of the function (integral of TNFW), which are the deflection angles
@@ -102,9 +101,7 @@ class TNFW(LensProfileBase):
         :param center_y: center of halo (in angular units)
         :return: deflection angle in x, deflection angle in y
         """
-        rho0_input = self._alpha2rho0(alpha_Rs=alpha_Rs, Rs=Rs)
-        #if Rs < 0.0000001:
-        #    Rs = 0.0000001
+        rho0_input = self.alpha2rho0(alpha_Rs=alpha_Rs, Rs=Rs)
         x_ = x - center_x
         y_ = y - center_y
         R = np.sqrt(x_ ** 2 + y_ ** 2)
@@ -127,9 +124,7 @@ class TNFW(LensProfileBase):
         :return: Hessian matrix of function d^2f/dx^2, d^f/dy^2, d^2/dxdy
         """
 
[…]
