### 2021-06-04 d4f1428
added TNFW_ELLIPSE profile - not tested yet
