+++ lenstronomy/LensModel/Profiles/tnfw_ellipse.py
@@ -1,0 +1,142 @@
+__author__ = 'sibirrer'
+
+import numpy as np
+from lenstronomy.LensModel.Profiles.tnfw import TNFW
+import lenstronomy.Util.param_util as param_util
+from lenstronomy.LensModel.Profiles.base_profile import LensProfileBase
+
+__all__ = ['TNFW_ELLIPSE']
+
+
+class TNFW_ELLIPSE(LensProfileBase):
+    """
+    this class contains functions concerning the truncated NFW profile with an ellipticity defined in the potential
+    parameterization of alpha_Rs, Rs and r_trunc is the same as for the spherical NFW profile
+
+    from Glose & Kneib: https://cds.cern.ch/record/529584/files/0112138.pdf
+
+    relation are: R_200 = c * Rs
+    """
+    profile_name = 'TNFW_ELLIPSE'
+    param_names = ['Rs', 'alpha_Rs', 'r_trunc', 'e1', 'e2', 'center_x', 'center_y']
+    lower_limit_default = {'Rs': 0, 'alpha_Rs': 0, 'r_trunc': 0, 'e1': -0.5, 'e2': -0.5, 'center_x': -100, 'center_y': -100}
+    upper_limit_default = {'Rs': 100, 'alpha_Rs': 10, 'r_trunc': 100, 'e1': 0.5, 'e2': 0.5, 'center_x': 100, 'center_y': 100}
+
+    def __init__(self):
+        """
+
+        """
+        self.tnfw = TNFW()
+        self._diff = 0.0000000001
+        super(TNFW_ELLIPSE, self).__init__()
+
+    def function(self, x, y, Rs, alpha_Rs, r_trunc, e1, e2, center_x=0, center_y=0):
+        """
+        returns elliptically distorted NFW lensing potential
+
+        :param x: angular position (normally in units of arc seconds)
+        :param y: angular position (normally in units of arc seconds)
+        :param Rs: turn over point in the slope of the NFW profile in angular unit
+        :param alpha_Rs: deflection (angular units) at projected Rs
+        :param r_trunc: truncation radius
+        :param e1: eccentricity component in x-direction
+        :param e2: eccentricity component in y-direction
+        :param center_x: center of halo (in angular units)
+        :param center_y: center of halo (in angular units)
+        :return: lensing potential
+        """
+        x_, y_ = param_util.transform_e1e2_square_average(x, y, e1, e2, center_x, center_y)
+        R_ = np.sqrt(x_**2 + y_**2)
+        rho0_input = self.tnfw.alpha2rho0(alpha_Rs=alpha_Rs, Rs=Rs)
+        if Rs < 0.0000001:
+            Rs = 0.0000001
+        f_ = self.tnfw.nfwPot(R_, Rs, rho0_input, r_trunc)
+        return f_
+
+    def derivatives(self, x, y, Rs, alpha_Rs, r_trunc, e1, e2, center_x=0, center_y=0):
+        """
+        returns df/dx and df/dy of the function, calculated as an elliptically distorted deflection angle of the
+        spherical NFW profile
+
+        :param x: angular position (normally in units of arc seconds)
+        :param y: angular position (normally in units of arc seconds)
+        :param Rs: turn over point in the slope of the NFW profile in angular unit
+        :param alpha_Rs: deflection (angular units) at projected Rs
+        :param r_trunc: truncation radius
+        :param e1: eccentricity component in x-direction
+        :param e2: eccentricity component in y-direction
+        :param center_x: center of halo (in angular units)
+        :param center_y: center of halo (in angular units)
[…]
