### 2021-06-05 d89d7a6
minor test improvements
