+++ Paper/paper.md
@@ -187,6 +187,7 @@ automated pipelines for gravitational lensing reconstruction [[dolphin](https://
 cluster source reconstruction and local perturbative lens modeling [[lenstruction](https://github.com/ylilan/lenstruction); @Yang:2020], 
 enhancement in large-scale structure imaging survey simulations [[DESC SLSprinkler](https://github.com/LSSTDESC/SLSprinkler); @LSSTDESC:2021], 
 rendering of sub-halos and line-of-sight halos [[pyHalo](https://github.com/dangilman/pyHalo); @Gilman:2020], 
+galaxy morphology analysis [[galight](https://github.com/dartoon/galight); @Ding:2020],
 and hierarchical analyses to measure the Hubble constant [[hierArc](https://github.com/sibirrer/hierarc); @Birrer:2020tdcosmoiv].
 With the rise in popularity and the promises in dealing with ever complex data problems with fast deep-learning methods, 
 dedicated tools for simulating large datasets for applying such methods to strong gravitational lensing [[deeplenstronomy](https://github.com/deepskies/deeplenstronomy); @Morgan:2021], [[baobab](https://github.com/jiwoncpark/baobab); @Park:2021], 
