### 2021-06-08 f641d5f
ready for release 1.8.2 (to be accepted JOSS paper)
