+++ lenstronomy/__init__.py
@@ -1,6 +1,6 @@
 __author__ = 'Simon Birrer'
 __email__ = 'sibirrer@gmail.com'
-__version__ = '1.8.1'
+__version__ = '1.8.2'
 __credits__ = 'ETH Zurich, UCLA, Stanford'
 
 from .Util.package_util import short, laconic
