### 2021-06-08 #3283 @sibirrer
> * [ ]  Make a tagged release of your software, and list the version tag of the archived version here.
tag: 1.8.2_JOSS
