### 2021-06-08 #3283 @danielskatz
I note the zenodo DOI isn't yet working, but I see it listed in https://zenodo.org/record/4913700#.YL-vNi1h131 so will proceed for now
