### 2021-06-08 #3283 @sibirrer
@danielskatz Yes, can do that. Do I need to regenerate a Zenodo DOI and GitHub release or can I just add it to the branch on GitHub?
