### 2021-06-08 #3283 @danielskatz
I also have suggested some small edits in the paper in https://github.com/sibirrer/lenstronomy/pull/261 - please merge this, or let me know what you disagree with.
