+++ Paper/paper.bib
@@ -165,21 +165,26 @@ archivePrefix = {arXiv},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
+
 @ARTICLE{Ding:2021transient,
        author = {{Ding}, Xuheng and {Liao}, Kai and {Birrer}, Simon and {Shajib}, Anowar J. and {Treu}, Tommaso and {Yang}, Lilan},
-        title = "{Improved time-delay lens modelling and $H_0$ inference with transient sources}",
-      journal = {arXiv e-prints},
-     keywords = {Astrophysics - Cosmology and Nongalactic Astrophysics, Astrophysics - Astrophysics of Galaxies},
+        title = "{Improved time-delay lens modelling and H$_{0}$ inference with transient sources}",
+      journal = {\mnras},
+     keywords = {gravitational lensing: strong, methods: data analysis, cosmological parameters, Astrophysics - Cosmology and Nongalactic Astrophysics, Astrophysics - Astrophysics of Galaxies},
          year = 2021,
-        month = mar,
-          eid = {arXiv:2103.08609},
-        pages = {arXiv:2103.08609},
+        month = jul,
+       volume = {504},
+       number = {4},
+        pages = {5621-5628},
+          doi = {10.1093/mnras/stab1240},
 archivePrefix = {arXiv},
        eprint = {2103.08609},
  primaryClass = {astro-ph.CO},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2021arXiv210308609D},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2021MNRAS.504.5621D},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
+
 
 @ARTICLE{Pagano:2020,
        author = {{Pagano}, G. and {Hannuksela}, O.~A. and {Li}, T.~G.~F.},
