### 2021-06-08 #3283 @sibirrer
Thank you @danielskatz ! I added the DOI and merged the JOSS_paper branch to main. So you can trigger either from the JOSS_paper branch or main branch.
