### 2021-06-08 #3283 @sibirrer
> I'm going to leave this open until the DOI resolves, which hopefully will be soon and normally would have already happened, but the internet was broken earlier today and might still be catching up

I had also some issues with automatic zenodo pushes from GitHub and needed to do it manually. Hope it gets resolved but it's not urgent. Thank you very much!
