### 2021-06-08 b7760de
delete draft-pdf.yml file to make github actions not crash
