### 2021-06-08 3ff2fa7
updated attribution with new JOSS paper
