### 2021-06-08 ed19aad
setup.py updated to version 1.8.2
