+++ setup.py
@@ -46,7 +46,7 @@ PACKAGE_PATH = os.path.abspath(os.path.join(__file__, os.pardir))
 
 setup(
     name='lenstronomy',
-    version='1.8.1',
+    version='1.8.2',
     description='Strong lens modeling package.',
     long_description=desc,
     author='Simon Birrer',
