### 2021-06-08 #3283 @sibirrer
something is weird with the DOI, earlier and later announcements by JOSS did result in a link (the twitter announcement is kind of lost without the DOI link). Is there a way to re-trigger the DOI generation @danielskatz?
